requirements.txt
fastapi==0.111.0
python-dotenv==1.0.0
uvicorn==0.30.1
requests==2.25.1
ua-parser==0.18.0
validators==0.28.3